requirements.txt
flask
pyyaml
mwoauth
requests